Global values
Fixed bug with global values. Now the method `makeKeyEvent' takes the
arguments `newAttack', `newRelease' etc. rather than `attack',
`release', etc. The `playNote' and `playMIDINote' methods take the
`attack', `release' etc arguments.

The `makeKeyEvent' checks if the value is nil, and if so uses the
global values. This way, I can have a global value for all parameters
that doesn't get changed when playing a key with a new value.